pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "flag-forge"
version = "0.1.0"
description = "Git-centric feature flag system with local evaluation, SDK, and stale flag detection"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [{name = "Flag Forge Contributors"}]
keywords = ["feature-flags", "feature-toggles", "rollout", "experimentation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.scripts]
flag-forge = "flag_forge.cli:cli"

[project.urls]
Homepage = "https://github.com/flag-forge/flag-forge"
Issues = "https://github.com/flag-forge/flag-forge/issues"
